app/requirements.txt
asyncio-hn
sanic
lru-dict